I have created the 2 different styles that i thought of.
